requirements.txt
asyncio==3.4.3
chainlit==1.1.202
httpx==0.27.0
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-experimental==0.0.60
langchain-nvidia-ai-endpoints==0.1.2
langchain-text-splitters==0.2.1
langgraph==0.0.66
openai==1.34.0
pandas==2.1.4
python-dotenv==1.0.1
tabulate==0.9.0